pre-interim presentation styling (#43)
