+++ client/src/components/reviews/productBreakdown/ProductBreakdown.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const ProductBreakdown = () => {
+  // Add productBreakdown prop once it exists
+  return (
+    <div data-testid="productBreakdown">
+      <td>
+        <tr>
+          <td>Size: |---------^--| </td>
+        </tr>
+        <tr>
+          <td>Comfort:|---------^--| </td>
+        </tr>
+      </td>
+    </div>
+  );
+};
+
+export default ProductBreakdown;
