+++ client/src/components/reviews/productBreakdown/ProductBreakdown.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import ProductBreakdown from './ProductBreakdown.js';
+import ProductBreakdown from './ProductBreakdown';
 
 it('renders learn react link', () => {
   const { getByTestId } = render(<ProductBreakdown />);
