+++ client/src/components/reviews/reviewList/ReviewList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import ReviewList from './ReviewList.js';
+import ReviewList from './ReviewList';
 
 it('renders the Review List container', () => {
   const { getByTestId } = render(<ReviewList />);
