+++ client/src/components/reviews/reviewList/ReviewTile.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 const ReviewTile = ({ summary, body, reviewerName, date }) => {
   return (
-    <td data-testid="reviewTile">
+    <td className="review-card" data-testid="reviewTile">
       <p>{summary}</p>
       <p>{body}</p>
       <p>{reviewerName}</p>
