+++ client/src/components/reviews/reviewList/ReviewTile.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import ReviewTile from './ReviewTile.js';
+import ReviewTile from './ReviewTile';
 
 it('renders one or more Review Tiles', () => {
   const { getByTestId } = render(<ReviewTile />);
