+++ client/src/styles/Reviews.css
@@ -1,0 +1,9 @@
+.review-card {
+  border: 1px solid black;
+  position: relative;
+  margin-bottom: 25px;
+}
+
+.review-breakdowns {
+  valign: top;
+}
